+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to
 - Remove `owner` and `is_main` in CreditCard model permanently
 - Remove `is_active` on order group client serializer
 
+## Fixed
+
+- BO: generate certificates for orders with product of type certificate
+
 ## [2.16.0] - 2025-02-13
 
 ### Added